+++ gui.h
@@ -4,7 +4,7 @@ void gui_init();
 void gui_exit(int);
 void gui_log(const char *s);
 char *gui_input(char *s);
-void gui_display(int row , float cur , float quo);
+void gui_display(int row , int cur , int quo);
 
 #define UIInit gui_init
 #define UIExit gui_exit
